# docs/openapi_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Content Automation ERP API",
    "description": "Production-grade API-First Control Plane for YouTube Automation & Durable Workflows",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ideas/elements": {
      "get": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "List Elements",
        "operationId": "list_elements_ideas_elements_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElementSchema"
                  },
                  "title": "Response List Elements Ideas Elements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/elements/{element_id}/ideas": {
      "get": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "List Ideas For Element",
        "operationId": "list_ideas_for_element_ideas_elements__element_id__ideas_get",
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Element Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdeaSchema"
                  },
                  "title": "Response List Ideas For Element Ideas Elements  Element Id  Ideas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ideas/{idea_id}/gates": {
      "get": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "Get Idea Gates",
        "operationId": "get_idea_gates_ideas__idea_id__gates_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/{idea_id}": {
      "get": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "Get Idea Details",
        "operationId": "get_idea_details_ideas__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas": {
      "post": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "Create Idea",
        "operationId": "create_idea_ideas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdeaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/elements/{element_id}/generate": {
      "post": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "Generate Ideas For Element",
        "operationId": "generate_ideas_for_element_ideas_elements__element_id__generate_post",
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Element Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateElementIdeasRequest",
                "default": {
                  "skip_browser": false,
                  "target_total": 10
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateElementIdeasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/categories/{category_id}/elements/generate": {
      "post": {
        "tags": [
          "Ideas & Taxonomy"
        ],
        "summary": "Generate Element From Category",
        "operationId": "generate_element_from_category_ideas_categories__category_id__elements_generate_post",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateElementRequest",
                "default": {
                  "skip_browser": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateElementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/packages/{idea_id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Get Package Manifest",
        "operationId": "get_package_manifest_packages__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageManifestSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{idea_id}/bundle": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Bundle Package",
        "operationId": "bundle_package_packages__idea_id__bundle_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          },
          {
            "name": "skip_browser",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Skip Browser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/audio/{idea_id}/status": {
      "get": {
        "tags": [
          "Audio & Speech"
        ],
        "summary": "Get Audio Status",
        "operationId": "get_audio_status_audio__idea_id__status_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/{idea_id}/generate": {
      "post": {
        "tags": [
          "Audio & Speech"
        ],
        "summary": "Generate Voiceover",
        "operationId": "generate_voiceover_audio__idea_id__generate_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAudioRequest",
                "default": {
                  "voice_id": "male_narrator"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateAudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioStatusResponse": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "has_voiceover": {
            "type": "boolean",
            "title": "Has Voiceover"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "enabled",
          "has_voiceover"
        ],
        "title": "AudioStatusResponse"
      },
...
      "CreateIdeaRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "default": 1
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "default": "Impossible Giant Machine"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "element_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Id"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateIdeaRequest"
      },
...
      "ElementSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "group_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uuid",
          "name"
        ],
        "title": "ElementSchema"
      },
...
      "GenerateAudioRequest": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id",
            "default": "male_narrator"
          }
        },
        "type": "object",
        "title": "GenerateAudioRequest"
      },
      "GenerateAudioResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "voice_id"
        ],
        "title": "GenerateAudioResponse"
      },
      "GenerateElementIdeasRequest": {
        "properties": {
          "skip_browser": {
            "type": "boolean",
            "title": "Skip Browser",
            "default": false
          },
          "target_total": {
            "type": "integer",
            "title": "Target Total",
            "default": 10
          }
        },
        "type": "object",
        "title": "GenerateElementIdeasRequest"
      },
      "GenerateElementIdeasResponse": {
        "properties": {
          "element_id": {
            "type": "integer",
            "title": "Element Id"
          },
          "element_name": {
            "type": "string",
            "title": "Element Name"
          },
          "generated_count": {
            "type": "integer",
            "title": "Generated Count"
          },
          "ideas": {
            "items": {
              "$ref": "#/components/schemas/IdeaSchema"
            },
            "type": "array",
            "title": "Ideas"
          }
        },
        "type": "object",
        "required": [
          "element_id",
          "element_name",
          "generated_count",
          "ideas"
        ],
        "title": "GenerateElementIdeasResponse"
      },
      "GenerateElementRequest": {
        "properties": {
          "skip_browser": {
            "type": "boolean",
            "title": "Skip Browser",
            "default": false
          }
        },
        "type": "object",
        "title": "GenerateElementRequest"
      },
      "GenerateElementResponse": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "element": {
            "$ref": "#/components/schemas/ElementSchema"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "element"
        ],
        "title": "GenerateElementResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdeaSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "raw_idea": {
            "type": "string",
            "title": "Raw Idea"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "element_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Id"
          },
          "element_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uuid",
          "title",
          "raw_idea",
          "status"
        ],
        "title": "IdeaSchema"
      },
...
      "PackageManifestSchema": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "folder_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Path"
          },
          "folder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Name"
          },
          "has_video": {
            "type": "boolean",
            "title": "Has Video"
          },
          "video_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Path"
          },
          "has_prompt_json": {
            "type": "boolean",
            "title": "Has Prompt Json"
          },
          "has_seo_json": {
            "type": "boolean",
            "title": "Has Seo Json"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "exists",
          "has_video",
          "has_prompt_json",
          "has_seo_json",
          "complete"
        ],
        "title": "PackageManifestSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}